requirements.txt
pydantic~=2.5
openai~=1.3
anthropic~=0.7
google-generativeai~=0.3
python-dotenv~=1.0
PyMuPDF~=1.23
pytesseract~=0.3
Pillow~=10.1
typer~=0.9